add support for user-supplied random number generator and refactor
